requirements.txt
Django==4.2.7
Pillow==10.1.0
openai==1.30.1
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0